requirements.txt
djangorestframework              3.12.4
Django                           3.1.5
Python                           3.8.5